* refactor: entities/model 추가 및 파일 경로 변경
